refactor(exception): statusCode 결정 관련 로직 이관
- errorCode를 통해 statusCode를 추론하는 로직을 service 로 이관하였습니다.
